ultima actualiacion, no anda agregar cuadrilla

+++ CommonSolution/DTO/DtoCuadrilla.cs
@@ -23,7 +23,7 @@ namespace CommonSolution.Dto
         [DisplayName("cantidadPeones")]
         [StringLength(100, ErrorMessage = "El {0} no puede superar {1}")]
         [Required(ErrorMessage = "El {0} es requerido")]
-        public short cantidadPeones { get; set; }
+        public int cantidadPeones { get; set; }
 
         public long idZona;
         public List<DtoReclamo> colReclamos;
